NEW Configure logging for syslog automatically
So all Silverstripe.Cloud projects do not miss critical errors in
greylog. As per platform documentation, there are 2 required modules
(dynamodb and crontask), and syslog logging should be set up at the
application level. The project currently has the required modules, but
misses the logging setup. This means critical errors in production
environments do not appear in the logs, and this is bad. Logging is
better than bad, it's good.

The Silverstripe vendored Common Web Platform on the other hand already
contains this logging information in the required cwp/cwp-core module,
so this configuration is skipped if that module exists - in order to
avoid a double-up of log entries.