Update windows yarn build copy dist files check
Windows does support a feature similar to symbolic links, and most modern systems are smart enough to know to use it when a symlink is asked for. This makes the "do a copy on Windows" no longer a default true, and we must check if we need to copy or not. In some cases a copy may still be needed, so simply removing the existing check is insufficient - an edge case but a case none the less.